refactor: split thread rail component
