+++ packages/playground-ui/src/ds/components/MessageScroller/index.ts
@@ -1,4 +1,2 @@
 export * from './message-scroller';
 export * from './message-scroller-hooks';
-export * from './thread-rail';
-export * from './thread-rail-turns';
