+++ packages/playground/src/lib/ai-ui/thread.tsx
@@ -3,17 +3,15 @@ import { Avatar } from '@mastra/playground-ui/components/Avatar';
 import { Button } from '@mastra/playground-ui/components/Button';
 import { ButtonsGroup } from '@mastra/playground-ui/components/ButtonsGroup';
 import {
-  buildThreadRailTurns,
-  getClientMessageKey,
   MessageScrollerContent,
   MessageScrollerItem,
   MessageScrollerProvider,
   MessageScrollerViewport,
-  ThreadRail,
 } from '@mastra/playground-ui/components/MessageScroller';
-import type { ThreadRailTurn } from '@mastra/playground-ui/components/MessageScroller';
 import { PendingIndicator } from '@mastra/playground-ui/components/PendingIndicator';
 import { ScrollArea } from '@mastra/playground-ui/components/ScrollArea';
+import { buildThreadRailTurns, getClientMessageKey, ThreadRail } from '@mastra/playground-ui/components/ThreadRail';
+import type { ThreadRailTurn } from '@mastra/playground-ui/components/ThreadRail';
 import { useAutoscroll } from '@mastra/playground-ui/hooks/use-autoscroll';
 import { cn } from '@mastra/playground-ui/utils/cn';
 import type { MessageFactoryPart } from '@mastra/react';
